requirements.txt
# requirements.txt

fastapi==0.103.1
uvicorn[standard]==0.23.1
sqlalchemy==2.0.21
alembic==1.11.1
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
python-multipart==0.0.6
pydantic==2.2.1
pydantic-settings==2.0.3
redis==4.5.1
